Split platform-specific code into per-platform files.
