+++ src/util/CMakeLists.txt
@@ -4,9 +4,17 @@ target_sources(zeek-agent
     PRIVATE
         ascii-table.cc
         helpers.cc
-        platform.cc
+        platform.testing.cc
         result.cc
 )
+
+if ( HAVE_DARWIN )
+    target_sources(zeek-agent PRIVATE platform.posix.cc platform.darwin.cc)
+endif ()
+
+if ( HAVE_LINUX )
+    target_sources(zeek-agent PRIVATE platform.posix.cc platform.linux.cc)
+endif ()
 
 if ( HAVE_WINDOWS )
     target_sources(zeek-agent PRIVATE platform.windows.cc)
